style: Rename component `ProductsFilters`
