+++ src/organisms/ProductsFilters/ProductsFilters.stories.tsx
@@ -1,13 +1,13 @@
 import { ComponentMeta, Story } from '@storybook/react'
 
-import { ProductsFilters, ProductsFiltersProps } from './ProductsSearchFilter'
+import { ProductsFilters, ProductsFiltersProps } from './ProductsFilters'
 
 const categories = ['All', 'Category 1', 'Category 2', 'Category 3']
 const subCategories = ['All', 'SubCategory 1', 'SubCategory 2', 'SubCategory 3']
 const brands = ['All', 'Brand 1', 'Brand 2', 'Brand 3']
 
 export default {
-  title: 'ProductsFilters',
+  title: 'organisms/ProductsFilters',
   component: ProductsFilters,
 } as ComponentMeta<typeof ProductsFilters>
